ai-model/requirements.txt
# AI model training (Python 3.10–3.12). Install in venv or use backend venv.
tensorflow>=2.15.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
